Cut the log overlay's per-event cost and split logger.rs into modules
The ring buffer layer allocated three strings per event on whatever thread
logged, including the video pump, and kept capturing while the overlay was
frozen and nobody was reading it.
