+++ src/logger/launch.rs
@@ -1,0 +1,46 @@
+//! Launch parameters SAM hands the app as argv[1].
+use std::sync::OnceLock;
+
+use serde::Deserialize;
+
+use crate::services::store::LogLevelOverride;
+
+/// argv[1] shape from SAM; both fields optional (no error if missing).
+#[derive(Deserialize, Default)]
+struct LaunchParams {
+    telemetry: Option<String>,
+    telemetry_level: Option<String>,
+}
+
+/// Cache launch params once; argv doesn't change over process lifetime.
+fn launch_params() -> &'static LaunchParams {
+    static PARAMS: OnceLock<LaunchParams> = OnceLock::new();
+    PARAMS.get_or_init(|| {
+        std::env::args()
+            .nth(1)
+            .and_then(|raw| serde_json::from_str(&raw).ok())
+            .unwrap_or_default()
+    })
+}
+
+/// Where to stream logs instead of writing them to disk; `None` means file.
+pub(super) fn telemetry_addr() -> Option<&'static str> {
+    launch_params().telemetry.as_deref().filter(|s| !s.is_empty())
+}
+
+/// Launch-time log level override from `TELEMETRY_LEVEL` env var.
+/// Folded into settings so Diagnostics can display it. `None` leaves persisted level.
+pub fn launch_level_override() -> Option<LogLevelOverride> {
+    match launch_params()
+        .telemetry_level
+        .as_deref()?
+        .to_ascii_lowercase()
+        .as_str()
+    {
+        "debug" => Some(LogLevelOverride::Debug),
+        "info" => Some(LogLevelOverride::Info),
+        "warn" => Some(LogLevelOverride::Warn),
+        "error" => Some(LogLevelOverride::Error),
+        _ => None,
+    }
+}
